tests/wpt/tests/import-maps/acquiring/modulepreload-link-header.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const t = async_test(
  'With modulepreload link header, import maps should fire error events');
const log = [];
</script>
<script type="importmap" onerror="t.done()">
{
  "imports": {
    "../resources/log.js?pipe=sub&name=A": "../resources/log.js?pipe=sub&name=B"
  }
}
</script>
<script>
promise_test(() => {
  return import("../resources/log.js?pipe=sub&name=A")
    .then(() => assert_array_equals(log, ["log:A"]))
  },
  'With modulepreload link header, import maps are not effective');
</script>
